+++ apps/web/src/app/api/stripe/portal/payment-methods/route.ts
@@ -1,0 +1,3 @@
+import { createStripePortalSessionHandler } from "../helpers/create-portal-session";
+
+export const GET = createStripePortalSessionHandler("/payment-methods");
